Mapping code for Part2
use ultrosonic to detect object and convert distance to coordinates. Then transform the coordinates to a numpy array where the object is 1 and no object is 0.
I had issue with this code. I supposed the y coordinates should be always positive if the car positioned at [0,0] whenever it starts to scan. But my code can return negative y coordinate.